requirements.txt
google-adk==1.17.0
google-genai==1.50.0
httpx==0.28.1
requests==2.32.5
fastapi==0.121.1
uvicorn==0.38.0
starlette==0.49.3
sse-starlette==3.0.3
pydantic==2.11.10
pydantic-settings==2.12.0
pydantic_core==2.33.2
anyio==4.11.0
protobuf==6.33.0
typing_extensions==4.15.0
python-dotenv==1.2.1
orjson==3.11.4
gradio==5.49.1
gradio_client==1.13.3
